# apidocs.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Toko Makanan",
      "description": "Daftar Makanan Restaurant",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "tags": [
      {
        "name": "Makanan",
        "description": "Makanan"
      }
    ],
    "paths": {
      "/makanan": {
        "get": {
          "summary": "Get all foods",
          "tags": ["Makanan"],
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "example": [
                    {
                      "id": 1,
                      "nama_makanan": "Nasi Goreng",
                      "harga": 15000
                    },
                    {
                      "id": 2,
                      "nama_makanan": "Ayam Goreng",
                      "harga": 20000
                    }
                  ]
                }
              }
            }
          }
        },
        "post": {
          "summary": "Add a new food",
          "tags": ["Makanan"],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nama_makanan": {
                      "type": "string"
                    },
                    "harga": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Food created successfully",
              "content": {
                "application/json": {
                  "example": {
                    "id": 3,
                    "nama_makanan": "Mie Goreng",
                    "harga": 12000
                  }
                }
              }
            }
          }
        }
      },
      "/makanan/update/{id}": {
        "put": {
          "summary": "Update a food by ID",
          "tags": ["Makanan"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the food to update",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nama_makanan": {
                      "type": "string"
                    },
                    "harga": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Food updated successfully",
              "content": {
                "application/json": {
                  "example": {
                    "id": 3,
                    "nama_makanan": "Mie Goreng",
                    "harga": 15000
                  }
                }
              }
            }
          }
        }
      },
      "/makanan/delete/{id}": {
        "delete": {
          "summary": "Delete a food by ID",
          "tags": ["Makanan"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the food to delete",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Food deleted successfully",
              "content": {
                "application/json": {
                  "example": {
                    "message": "Makanan berhasil dihapus"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
  